Create README.md
good afternoon, I would like to propose some improvements for your wallet.
As you know there are initiatives such as cardwallet that intends to use debit cards to spend the coins in the wallet. This represents an advantage and also a problem due to the KYC identification processes. That is why I think that a solution could be to pass our cryptocoins to fiat and then be able to fill or load a prepaid card with a pin.
This card is nameless and can be spent anonymously. You could even earn money by issuing the physical cards, in addition to charging a small commission per load.
I think this could keep the privacy of people who want to spend their crypto anonymously after transferring it to fiat inside the wallet.
Please tell me what you think of the idea.